+++ changelog.md
@@ -1,6 +1,12 @@
 Changelog
 
-#### 2017-11-09 - Kent Huang
+### #2 2017-11-09 - Kent Huang
+- added singleFlight & singlePlane models
+- added csv data gitignore
+
+### #1 2017-11-09 - Kent Huang
 - added fleet.csv & flights.csv
 - added model files(CSV,entity,etc.) in application/core 
 - loaded csv into Fleet and Flight collection models
+
+